:recycle: 코멘트 생성 API 스펙 변경으로 인한 Controller, REST Docs 변경 (#263)
