+++ application/src/docs/asciidoc/comment/community-comments.adoc
@@ -1,0 +1,26 @@
+=== /community-comments
+
+==== 커뮤니티 코멘트 조회
+
+.Request
+include::{snippets}/get-community-comments/http-request.adoc[]
+.Query Parameters
+include::{snippets}/get-community-comments/query-parameters.adoc[]
+.Response
+include::{snippets}/get-community-comments/http-response.adoc[]
+.Response Fields
+include::{snippets}/get-community-comments/response-fields.adoc[]
+
+==== 커뮤니티 코멘트 작성
+
+.Request
+include::{snippets}/create-community-comment/http-request.adoc[]
+.Request Headers
+include::{snippets}/create-community-comment/request-headers.adoc[]
+.Request Fields
+include::{snippets}/create-community-comment/request-fields.adoc[]
+.Response
+include::{snippets}/create-community-comment/http-response.adoc[]
+.Response Fields
+include::{snippets}/create-community-comment/response-fields.adoc[]
+
